order_by_relevance param added to whooshee_search method together with a unittest for it
order_by_relevance=n means:
n < 0  : return everything ordered
n > 0  : return first n results ordered
n == 0 : return everything unordered

We introduced this parameter to be able to limit the resulting query size
because the order_by_relevance is implemented by potentially huge CASE
statement.

The alternative was to use ORDER BY FIELD(<fieldname>, 1, 5, ...)
directive but this is not currently implemented in netheir postgresql
or sqlite (it is in MySQL though).

Also e.g. sqlite puts limit on number of variables in a query (given by
SQLITE_LIMIT_VARIABLE_NUMBER being set to 999 by default). By limiting
the number of search results for which the order is ensured, we also
limit the number of the variables, which might make an invalid query valid.

This can (and perhaps should from the user perspective) also be achieved
by using the `limit` option of whooshee_search method so order_by_relevance
represents just an extra degree of freedom in the search configuration.