tests: compat with flask-sqlalchemy v3
The 'create_all()' and 'drop_all()' methods no longer have the 'app='
argument.  Therefore we have to use 'with app.app_context()' wrapper, at
least when there are multiple apps in action.

Otherwise, it is possible to take the app context on a per-test-case
basis, using the pattern:

    setUp():
        self.ctx = self.app.app_context()
        self.ctx.push()

    tearDown():
        self.ctx.pop()

The changes done in this commit still keep compatibility with
flask-sqlalchemy v2.

Fixes: #75
Closes: #74
Closes: #73
Closes: #78